Refactor Header: Extract search, add mobile menu with animations
